feat(1635): Move types to separated file
