+++ src/core/types/transaction.types.ts
@@ -1,0 +1,13 @@
+import type { Transaction } from '@hashgraph/sdk';
+
+export interface BaseBuildTransactionResult {
+  transaction: Transaction;
+}
+
+export interface BaseSignTransactionResult {
+  signedTransaction: Transaction;
+}
+
+export interface BaseNormalizedParams {
+  keyRefIds: string[];
+}
